+++ android/app/src/main/java/com/noxplay/noxplayer/MainActivity.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.noxplay.noxplayer;
 
 import android.os.Bundle;
