+++ android/app/src/release/java/com/example/ReactNativeFlipper.java
@@ -4,7 +4,7 @@
  * <p>This source code is licensed under the MIT license found in the LICENSE file in the root
  * directory of this source tree.
  */
-package com.example;
+package com.noxplay.noxplayer;
 
 import android.content.Context;
 import com.facebook.react.ReactInstanceManager;
